Add guess functionality

+++ src/main/java/com/poleszak/GuessGame/dto/GameBestTenDto.java
@@ -1,4 +1,4 @@
 package com.poleszak.GuessGame.dto;
 
-public record GameBestTenDto(Long id, int numberOfAttempts, int gameTimeInSeconds) {
+public record GameBestTenDto(Long id, int numberOfAttempts, Long gameTimeInSeconds) {
 }
